Checkpoint that shows how the graph info can be used.
Since the details (say Module/ModuleCall in nodeExpandModule) are private to terraform package, one can either use the public interfaces to get info, or to add helpers in the terraform package.

The current public interfaces expose some bits of detail, but it seems far not enough to reconstruct precise sources of module internals or calls (now source details are only available about references, but not about definitions, for example).

So one could add further public interfaces to expose the required data.

But, we could maybe just get back to the previous approach of directly accessing the root module, and traversing top-down, which would give us more details with less hassle.